Remove public references in common/
