+++ common/index.ts
@@ -6,6 +6,3 @@
 export * from './constants';
 export * from './interfaces';
 export * from './utils';
-export * from '../public/component_types';
-export * from '../public/utils';
-export * from '../public/pages/workflow_detail/utils';
