+++ public/component_types/index.ts
@@ -3,7 +3,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './interfaces';
 export * from './transformer';
 export * from './indexer';
 export * from './other';
