--- public/component_types/interfaces.ts
@@ -1,91 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { FormikValues } from 'formik';
-import { ObjectSchema } from 'yup';
-import { COMPONENT_CATEGORY, COMPONENT_CLASS } from '../utils';
-
-/**
- * ************ Types *************************
- */
-export type FieldType = 'string' | 'json' | 'select' | 'model';
-export type SelectType = 'model';
-export type FieldValue = string | {};
-export type ComponentFormValues = FormikValues;
-export type WorkspaceFormValues = {
-  [componentId: string]: ComponentFormValues;
-};
-export type WorkspaceSchemaObj = {
-  [componentId: string]: ObjectSchema<any, any, any>;
-};
-export type WorkspaceSchema = ObjectSchema<WorkspaceSchemaObj>;
-
-/**
- * Represents a single base class as an input handle for a component.
- * It may accept multiples of that class.
- */
-export interface IComponentInput {
-  id: string;
-  label: string;
-  baseClass: COMPONENT_CLASS;
-  acceptMultiple: boolean;
-}
-
-/**
- * An input field for a component. Specifies enough configuration for the
- * UI node to render it properly (help text, links, etc.)
- */
-export interface IComponentField {
-  label: string;
-  type: FieldType;
-  id: string;
-  value?: FieldValue;
-  placeholder?: string;
-  helpText?: string;
-  helpLink?: string;
-  selectType?: SelectType;
-}
-
-/**
- * Represents the list of base classes as a single output handle for
- * a component.
- */
-export interface IComponentOutput {
-  label: string;
-  baseClasses: COMPONENT_CLASS[];
-}
-
-/**
- * The base interface the components will implement.
- */
-export interface IComponent {
-  type: COMPONENT_CLASS;
-  label: string;
-  description: string;
-  // will be used for grouping together in the drag-and-drop component library
-  // and determining which flows the component can be drug into the workspace flows
-  categories: COMPONENT_CATEGORY[];
-  // determines if this component allows for new creation. this means to
-  // allow a "create" option on the UI component, as well as potentially
-  // include in the use case template construction ('provisioning' flow)
-  allowsCreation: boolean;
-  // the list of base classes that will be used in the component output
-  baseClasses?: COMPONENT_CLASS[];
-  inputs?: IComponentInput[];
-  fields?: IComponentField[];
-  // if the component supports creation, we will have a different set of input fields
-  // the user needs to fill out
-  createFields?: IComponentField[];
-  outputs?: IComponentOutput[];
-}
-
-/**
- * We need to include some extra instance-specific data to the ReactFlow component
- * to perform extra functionality, such as deleting the node from the ReactFlowInstance.
- */
-export interface IComponentData extends IComponent {
-  id: string;
-  selected?: boolean;
-}
