+++ public/component_types/other/query/neural_query.tsx
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { COMPONENT_CLASS } from '../../../utils';
+import { COMPONENT_CLASS } from '../../../../common';
 import { Query } from './query';
 
 /**
