+++ public/component_types/transformer/normalization_transformer.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { COMPONENT_CATEGORY, COMPONENT_CLASS } from '../../utils';
+import { COMPONENT_CATEGORY, COMPONENT_CLASS } from '../../../common';
 import { ResultsTransformer } from './results_transformer';
 
 /**
