+++ public/component_types/transformer/results_transformer.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { COMPONENT_CLASS } from '../../utils';
+import { COMPONENT_CLASS } from '../../../common';
 import { BaseTransformer } from './base_transformer';
 
 /**
